Organize better core/instance code in files
Split instance.Config stuff out of instance.go to a new config.go.

Split instance.Status stuff out of instance.go to a new status.go.

Split instance.Monitor leftover stuff out of instance.go to the
already existing monitor.go.
